04/kadai/requirements.txt
ffmpeg-python
mlx-whisper
pydub
numpy

